move TransactionStatusBadge to components/status folder
